move td64 tests from series tests; unxfail fixed
